All operations now require an explicit Connection
The netfs operations were callable in arbitrary order and the calls were
not aware of any underlying connection losses. This was leading to
various strange behaviour from the server's point of view, where
connections were dropped arbitrarily and re-initiated with the committal
of an empty transaction, for example.

This patch fixes all these issues by requiring an explicit
NetfsConnection object to perform *any* of the netfs operations